requirements/testing.txt
# Test dependencies for the labman project
# Install with: pip install -r requirements/testing.txt

# Core testing
pytest>=8.0
pytest-django>=4.8
pytest-cov>=5.0

# Code coverage
coverage[toml]>=7.0
